Make test harness run different parser for each subdir.
We create a `directory_checker_map` that maps
subdirectories to checkers. A checker is a function
that takes a string and returns an error if the parse was
unsuccessful.

We run these chekers for each subdirectory, and set the
return value of the overall tool to 0 or 1 based on the
success / failure of parsing.

We also print the output of each test case in CSV format,
anticipating post-processing for the paper