Refactor: Use Except to represent parse error.
instad of using `Option String \times String`, which has unclear
semantics, use `Except ParseError ParseOutput` to clearly state
what the function is supposed to do.

Also, see that we introducec the indirection of the type `AnyNonterminal.*`,
to call `toString` on it. Avoid this indirection by immediately calling
`toString` on the nonterminal we produce, thereby eliminating the need
for `AnyNonterminal`.

In general, if we introduce an inductive only to dispatch on a
typeclass, it is often possible to directly hold a reference to the
typeclass itself: https://softwareengineering.stackexchange.com/a/311833